moved trsp headers to the correct location
