+++ include/trsp/GraphDefinition.h
@@ -12,7 +12,7 @@
 #include <functional>
 
 
-#include "trsp.h"
+#include "c_types/trsp/trsp.h"
 
 // using namespace std;
